requirements.txt
pyautogui
easyocr
opencv-python
pillow
numpy
ollama
plyer
pygetwindow
